+++ src/main.tsx
@@ -1,13 +1,13 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import './index.css';
-import App from './App';
+import PublicRouter from './PublicRouter';
 import { I18nProvider } from './i18n';
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
     <I18nProvider>
-      <App />
+      <PublicRouter />
     </I18nProvider>
   </StrictMode>,
 );